### 2017-12-22 c8882ba
fix small bug in readGP
