### 2017-12-22 bb46913
remove debug statements
