+++ R/writeGP.R
@@ -110,10 +110,6 @@ readGP = function(file) {
   g$node.types   = strsplit(scan(con, what = character(), n = 1L, quiet = TRUE), ",")[[1L]]
   g$edge.types   = strsplit(scan(con, what = character(), n = 1L, quiet = TRUE), ",")[[1L]]
 
-  print(g$weight.types)
-  print(g$node.types)
-  print(g$edge.types)
-
   # import coordinates
   g$coordinates = as.matrix(read.table(con, sep = ",", nrows = n, header = FALSE, stringsAsFactors = FALSE))
 
