### 2018-01-03 251711f
compute adj.list on import (fixes issue with mcSSSP)
