### 2018-01-08 #520 @Kevin-Mattheus-Moerman
@gvegayon you reviewed related [work by this author](https://github.com/openjournals/joss-reviews/issues/374) in the past. Would you be able and willing to assist with review here too?
