### 2018-01-08 #520 @Kevin-Mattheus-Moerman
@gpeyre Is this submission of interest to you or your colleagues? It is an R-package on multi-criteria optimization problems on graphs, see also the summary in the [paper 📜 ](https://github.com/openjournals/joss-papers/blob/joss.00520/joss.00520/10.21105.joss.00520.pdf).
