### 2018-01-08 #528 @whedon
[REVIEW]: grapherator: A Modular Multi-Step Graph Generator
**Submitting author:** @jakobbossek (<a href="http://orcid.org/0000-0002-4121-4668">Jakob Bossek</a>)
**Repository:** <a href="https://github.com/jakobbossek/grapherator" target ="_blank">https://github.com/jakobbossek/grapherator</a>
**Version:** v1.0.0
