### 2018-02-05 #528 @gvegayon
Looks good in general. Now I'll take a look at the documentation and code.
