### 2018-02-12 jakobbossek/grapherator#27 @gvegayon
Review for JOSS
Regarding the review https://github.com/openjournals/joss-reviews/issues/528 I have the following comments:

- [x] In the vignette "Graph generation with grapherator", there are some references missing. In particular, the MOST paper, which is linked to, but not mentioned in the list of references.

- [x] To improve documentation, I suggest adding links from the `addEdges`, `addNodes`, and `addWeights` to their respective set of generator functions. A single link in the parameter description should work.

- [x] The tag `@title` shows in the name of `edgeGenerators` (see [here](https://github.com/jakobbossek/grapherator/blob/f5dabae851bc58dfe8be4314983f30137fd750bb/R/generators-edge.R#L1)). Removing the `@title` tag should be OK.

Other than that, everything looks OK to me. I was able to install the package from github and CRAN and compile/test it myself without any problem.
