### 2018-02-12 jakobbossek/grapherator#27 @jakobbossek
Thanks for the comments @gvegayon. Going to fix this stuff tomorrow.
