### 2018-02-13 580ed1e
fix JOSS review comments
