+++ vignettes/paper.bib
@@ -50,3 +50,18 @@
  address = {San Francisco, CA, USA},
 }
 
+@TECHREPORT{CJM08,
+  title={MOST – Multiple Objective Spanning Trees Repository Project},
+  author={Pedro Cardoso and M{\'a}rio Jesus and Alberto M{\'a}rquez},
+  year={2006},
+  institution={University of Algarve},
+  type={Technical Report},
+}
+
+@TECHREPORT{CJM05,
+  title={MONACO applied to the multiple criteria minimum spanning trees problem},
+  author={Pedro Cardoso and M{\'a}rio Jesus and Alberto M{\'a}rquez},
+  year={2008},
+  institution={University of Algarve},
+  type={Technical Report},
+}
