+++ man/weightGenerators.Rd
@@ -1,11 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/generators-weight.R
-\name{addWeightsConcave}
+\name{weightGenerators}
+\alias{weightGenerators}
 \alias{addWeightsConcave}
 \alias{addWeightsCorrelated}
 \alias{addWeightsDistance}
 \alias{addWeightsRandom}
-\title{@title Weight generators.}
+\title{Weight generators.}
 \usage{
 addWeightsConcave(graph, xhi = 10, nu = 20, M = 100, ...)
 
