### 2018-02-13 jakobbossek/grapherator#27 @jakobbossek
Fixed via 580ed1e
