### 2018-02-15 #528 @gvegayon
Everything looks good to me. I'm OK with the current state of the package.
