### 2018-02-19 #528 @Kevin-Mattheus-Moerman
@jakobbossek it looks like we are nearly there. Can you check your references for DOI's. Not all currently have DOI's, perhaps they are not available for all but please check.
