### 2018-02-19 cce8291
minor changes to bibfile
