+++ paper/paper.bib
@@ -34,6 +34,10 @@
   booktitle={{Proceedings of the 2017 IEEE Symposium Series on Computational Intelligence}},
   title={A Pareto-Beneficial Sub-Tree Mutation for the Multi-Criteria Minimum Spanning Tree Problem},
   year={2017},
+  pages = {3280--3287},
+  location = {Honolulu, Hawaii},
+  numpages = {8},
+  address = {Honolulu, Hawaii, USA}
 }
 
 @article{B17,
@@ -42,6 +46,7 @@
   doi = {10.21105/joss.00374},
   url = {https://doi.org/10.21105/joss.00374},
   year  = {2017},
+  isbn = {978-1-5386-4058-6},
   month = {sep},
   volume = {2},
   number = {17},
