### 2018-02-19 fe243af
fix typo in paper
