### 2018-02-19 #528 @jakobbossek
@Kevin-Mattheus-Moerman I tried hard to figure out DOI's for all the cited papers. Unfortuenately there are no DOI's available for most of them.
