### 2018-02-19 #528 @Kevin-Mattheus-Moerman
Okay thanks for checking. Did you fix the figure caption?
