+++ NEWS.md
@@ -1,3 +1,8 @@
+# grapherator 1.0.1
+
+* Fixed debug output in readGP
+* Fixed import bug in readGP
+
 # grapherator 1.0.0
 
 * first release to CRAN
