### 2018-02-19 #528 @jakobbossek
The DOI is 10.5281/zenodo.1175943
Thanks @Kevin-Mattheus-Moerman and @gvegayon 👏
